+++ packages/esm-appointments-app/src/aa-tbd/overlay.scss
@@ -1,4 +1,4 @@
-@import './style.scss';
+@import '../root.scss';
 
 .desktopOverlay {
   position: sticky;
