--- packages/esm-appointments-app/src/appointment-list/add-patient/add-patient.scss
@@ -1,55 +1,0 @@
-@import "../style.scss";
-
-.modalContent {
-  width: 100%;
-  background-color: #f4f4f4;
-}
-
-.modalHeader {
-  margin: 1rem;
-}
-
-.patientListList {
-  padding: 0 1rem;
-}
-
-.largeButtons {
-  padding-bottom: 2rem;
-}
-
-.checkbox {
-  padding-bottom: 0.875rem;
-}
-
-.pagination {
-  width: 100%;
-  overflow: hidden;
-}
-
-.paginationContainer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  position: relative;
-}
-
-.itemsCountDisplay {
-  position: absolute;
-  top: 0;
-  left: 1rem;
-  height: 3rem;
-  display: flex;
-  align-items: center;
-  color: #525252
-}
-
-.pagination > div:first-child {
-  visibility: hidden;
-  width: 0;
-}
-
-.buttonSet {
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-}
