--- packages/esm-appointments-app/src/appointment-list/constants.ts
@@ -1,5 +1,0 @@
-import { omrsOfflineCachingStrategyHttpHeaderName, OmrsOfflineHttpHeaders } from '@openmrs/esm-framework';
-
-export const cacheForOfflineHeaders: OmrsOfflineHttpHeaders = {
-  [omrsOfflineCachingStrategyHttpHeaderName]: 'network-first',
-};
