--- packages/esm-appointments-app/src/appointment-list/offline-patient-table/interactive-offline-patient-table.component.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import OfflinePatientTable from './offline-patient-table.component';
-
-const InteractiveOfflinePatientTable: React.FC = () => {
-  return <OfflinePatientTable isInteractive showHeader={false} />;
-};
-
-export default InteractiveOfflinePatientTable;
