--- packages/esm-appointments-app/src/appointment-list/offline-patient-table/last-updated-table-cell.scss
@@ -1,15 +1,0 @@
-@import "../style.scss";
-
-.errorIcon {
-  color: $danger;
-}
-
-.pendingIcon {
-  color: #ff832b;
-}
-
-.cellContainer {
-  display: flex;
-  align-items: center;
-  gap: $spacing-02;
-}
