--- packages/esm-appointments-app/src/appointment-list/offline-patient-table/overview-offline-patient-table.component.tsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import OfflinePatientTable from './offline-patient-table.component';
-
-const OverviewOfflinePatientTable: React.FC = () => {
-  return <OfflinePatientTable isInteractive={false} showHeader />;
-};
-
-export default OverviewOfflinePatientTable;
