+++ packages/esm-appointments-app/src/appointments-header/appointments-header.scss
