+++ packages/esm-appointments-app/src/appointments-metrics/appointment-metrics.resource.tsx
@@ -3,7 +3,7 @@ import useSWRImmutable from 'swr/immutable';
 import { FetchResponse, openmrsFetch } from '@openmrs/esm-framework';
 
 export function useMetrics() {
-  const metrics = { scheduled_appointments: 100, average_wait_time: 28, patients_waiting_for_service: 182 };
+  const metrics = { scheduled_appointments: 100, missed_appointments: 28, providers_available_today: 182 };
   const { data, error } = useSWR<{ data: { results: {} } }, Error>(`/ws/rest/v1/queue?`, openmrsFetch);
 
   return {
