+++ packages/esm-appointments-app/src/appointments-metrics/appointments-metrics.component.tsx
@@ -1,15 +1,14 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Dropdown, DataTableSkeleton } from 'carbon-components-react';
-import { useMetrics, useServices } from './queue-metrics.resource';
+import { DataTableSkeleton } from 'carbon-components-react';
+import { useMetrics } from './appointment-metrics.resource';
 import MetricsCard from './metrics-card.component';
 import MetricsHeader from './metrics-header.component';
-import styles from './clinic-metrics.scss';
+import styles from './appointments-metrics.scss';
 
-const ClinicMetrics: React.FC = () => {
+const AppointmentsMetrics: React.FC = () => {
   const { t } = useTranslation();
   const { metrics, isError, isLoading } = useMetrics();
-  const { services } = useServices();
 
   if (isLoading) {
     return <DataTableSkeleton role="progressbar" />;
@@ -20,18 +19,18 @@ const ClinicMetrics: React.FC = () => {
       <MetricsHeader />
       <div className={styles.cardContainer}>
         <MetricsCard
-          label={t('minutes', 'Minutes')}
-          value={metrics ? metrics.average_wait_time : 0}
+          label={t('appointments', 'Appointments')}
+          value={metrics ? metrics.scheduled_appointments : 0}
           headerLabel={t('scheduledAppointments', 'Scheduled appointments')}
         />
         <MetricsCard
-          label={t('patients', 'Patients')}
-          value={metrics ? metrics.scheduled_appointments : 0}
+          label={t('appointments', 'Appointments')}
+          value={metrics ? metrics.missed_appointments : 0}
           headerLabel={t('missedAppointments', 'Missed appts. today')}
         />
         <MetricsCard
-          label={t('patients', 'Patients')}
-          value={metrics ? metrics.patients_waiting_for_service : 0}
+          label={t('providers', 'Providers')}
+          value={metrics ? metrics.providers_available_today : 0}
           headerLabel={t('providersAvailableToday', 'Providers available today')}
         />
       </div>
@@ -39,4 +38,4 @@ const ClinicMetrics: React.FC = () => {
   );
 };
 
-export default ClinicMetrics;
+export default AppointmentsMetrics;
