+++ packages/esm-appointments-app/src/appointments-metrics/appointments-metrics.scss
