+++ packages/esm-appointments-app/src/appointments-metrics/metrics-card.scss
