+++ packages/esm-appointments-app/src/appointments-metrics/metrics-header.scss
