+++ packages/esm-appointments-app/src/dashboard.meta.ts
@@ -1,8 +1,8 @@
 import { Home32 } from '@carbon/icons-react';
 
-export const homeDashboardMeta = {
-  name: 'home',
-  slot: 'home-dashboard-slot',
+export const clinicalAppointmentsDashboardMeta = {
+  name: 'clinical-appointments',
+  slot: 'clinical-appointments-dashboard-slot',
   title: 'Clinical Appointments',
   renderIcon: Home32,
 };
