+++ packages/esm-appointments-app/src/patient-list-action-button.scss
