+++ packages/esm-appointments-app/src/tabs/appointments-table.scss
