+++ packages/esm-appointments-app/src/ui-components/create-edit-patient-list/create-edit-list.component.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback, SyntheticEvent, useEffect, useState } from 'react';
 import { Button, Dropdown, OnChangeData, TextArea, TextInput } from 'carbon-components-react';
-import Overlay from '../../overlay.component';
+import Overlay from '../../aa-tbd/overlay.component';
 import { useTranslation } from 'react-i18next';
 import styles from './create-edit-patient-list.scss';
 import { useLayoutType, showToast, useSessionUser } from '@openmrs/esm-framework';
