+++ packages/esm-appointments-app/translations/en.json
@@ -4,7 +4,6 @@
   "addNewAppointment": "Add New Appointment",
   "addPatientList": "Add patient to list",
   "addPatientToList": "Add patient to list",
-  "addToList": "Add to list",
   "advancedSearch": "Advanced search",
   "all": "All",
   "anotherVisitType": "Start another visit type",
@@ -18,7 +17,6 @@
   "configureList": "Configure your patient list using the fields below",
   "createdOn": "Created on",
   "createList": "Create list",
-  "createNewPatientList": "Create new patient list",
   "currentVisit": "Current visit",
   "dateOfBirth": "Date of birth",
   "dateTime": "Date & time",
@@ -34,7 +32,6 @@
   "encounterType": "Encounter type",
   "endVisit": "End visit",
   "error": "Error",
-  "errorAddPatientToList": "Patient not added to list",
   "errorCreatePatientListDescription": "Couldn't create patient list",
   "errorDeleteList": "Couldn't delete patient list",
   "errorUpdatePatientListDescription": "Couldn't update patient list",
@@ -44,7 +41,6 @@
   "futureScheduledVisits": " {count, plural, one{{count} visit} other{{count} visits}} scheduled for dates in the future",
   "futureScheduledVisits_plural": "",
   "identifier": "Identifier",
-  "items": "items",
   "lastName": "Last name",
   "lastVisit": "Last Visit",
   "lastVisitDate": "Date",
@@ -54,7 +50,6 @@
   "match": "Match",
   "medications": "Medications",
   "middleName": "Middle name",
-  "minutes": "Minutes",
   "missedAppointments": "Missed appts. today",
   "moreMetrics": "See more metrics",
   "name": "Name",
@@ -64,35 +59,14 @@
   "newPatientListHeader": "New patient list",
   "newPatientListNameLabel": "List name",
   "noEncountersFound": "No encounters found",
-  "noPatientListFound": "No patient list found",
-  "noPatientsToDisplay": "No patients to display",
+  "noAppointmentsToDisplay": "No appointments to display",
   "noPreviousVisitFound": "No previous visit found",
   "notes": "Notes",
   "notUrgent": "Not Urgent",
-  "offlinePatientsTableDeleteConfirmationModalCancel": "Cancel",
-  "offlinePatientsTableDeleteConfirmationModalConfirm": "Remove patients",
-  "offlinePatientsTableDeleteConfirmationModalContent": "Are you sure that you want to remove all selected patients from the offline list? The charts will no longer be available in offline mode.",
-  "offlinePatientsTableDeleteConfirmationModalTitle": "Remove offline patients",
-  "offlinePatientsTableHeaderAge": "Age",
-  "offlinePatientsTableHeaderGender": "Gender",
-  "offlinePatientsTableHeaderLastUpdated": "Last updated",
-  "offlinePatientsTableHeaderName": "Name",
-  "offlinePatientsTableLastUpdatedDownloading": "Downloading...",
-  "offlinePatientsTableLastUpdatedError": "error",
-  "offlinePatientsTableLastUpdatedErrors": "errors",
-  "offlinePatientsTableLastUpdatedNotYetSynchronized": "Not synchronized",
-  "offlinePatientsTableLastUpdatedOutdatedData": "Outdated data",
-  "offlinePatientsTableRemoveFromOfflineList": "Remove from list",
-  "offlinePatientsTableSearchLabel": "Search this list",
-  "offlinePatientsTableSearchPlaceholder": "Search this list",
-  "offlinePatientsTableTitle": "Offline patients",
-  "offlinePatientsTableUpdatePatient": "Update patient",
-  "offlinePatientsTableUpdatePatients": "Update patients",
   "openPatientList": "Add to list",
   "or": "or",
   "patientList": "Patient list",
   "patientListAppMenuLink": "Patient Lists",
-  "patientLists": "Patient Lists",
   "patients": "Patients",
   "personalDetails": "Personal Details",
   "phoneNumber": "Phone number",
@@ -101,14 +75,13 @@
   "primaryHelperText": "Search for a patient",
   "priority": "Priority",
   "provider": "Provider",
+  "providers": "Providers",
   "providersAvailableToday": "Providers available today",
   "recentScheduledVisits": "{count, plural, one{{count} visit} other{{count} visits}} scheduled for +/- 7 days",
   "recentScheduledVisits_plural": "",
   "scheduledAppointments": "Scheduled appointments",
   "search": "Search",
   "searchboxPlaceholder": "Search for a patient name or ID number",
-  "searchForAListToAddThisPatientTo": "Search for a list to add this patient to.",
-  "searchForList": "Search for a list",
   "searchThisList": "Search this list",
   "secondaryHelperText": "Type the patient's name or unique ID number",
   "selectCohortType": "Select patient list type",
@@ -117,12 +90,13 @@
   "showPatientsWaitingFor": "Show patients waiting for",
   "startDate": "Start Date",
   "status": "Status",
-  "successAddPatientToList": "Patient added to list",
   "successCreatedPatientList": "Created patient list",
   "successCreatedPatientListDescription": "Successfully created patient list",
-  "successfullyAdded": "Successfully added",
   "successUpdatePatientList": "Updated patient list",
   "successUpdatePatientListDescription": "Successfully updated patient list",
   "vitals": "Vitals",
+  "serviceType": "Service Type",
+  "location": "Location",
+  "view": "View",
   "waitTime": "Wait time (mins)"
 }
